add home page discover api

+++ app/src/main/java/com/tobery/personalmusic/entity/home/HomeDiscoverEntity.java
@@ -161,7 +161,7 @@ public class HomeDiscoverEntity {
                     @NoArgsConstructor
                     @Data
                     public static class ResourceExtInfoEntity {
-                        private int playCount;
+                        private long playCount;
                         private boolean highQuality;
                         private int specialType;
                     }
